compiling-rpms is out of date and redundant
The instructions on this page are outdated. Current (working) instructions for the extra step to make RPMS are on the main Building GlusterFS in this guide. Since this page is indexed it is probably better to leave it as a brief note to see the other page.

Experience -- I needed to build RPMS for Oracle EL 8.7 and followed this page but did not get installable packages, I followed the Building GlusterFS page and my only issue was that the gluster-selinux package is a dependency that has to be built from another repo. 

Improving my revision? Add a note on how to build the gluster-selinux rpm. That repo doesn't give instructions, I ended up cheating and using another distributions package for it.